Fixed doctor method with newer versions of the drchrono API
Newer versions of the drchrono API no longer pass _doctor_ as a URL, but
instead pass it as an ID. This is a breaking change if you are on older
versions of the API(v2014_08 or earlier). If you are using an older
version please use version 0.0.3 of this gem. This closes #3.